grouped Order concepts together. added error handling. gofmt
