+++ GetAccounts.go
@@ -2,23 +2,23 @@ package main
 
 // Get Accounts
 func GetAccounts() {
-  var err error
-  accounts, err = client.GetAccounts()
-  if err != nil {
-    println(err.Error())
-  }
+	var err error
+	accounts, err = client.GetAccounts()
+	if err != nil {
+		println(err.Error())
+	}
 
-  // Print Balances
-  for i, a := range accounts {
-    //println(a.Balance)
-    if a.Currency == "USD" {
-      usdIndex = i
-    }
-    if a.Currency == "BTC" {
-      btcIndex = i
-    }
-    if a.Currency == "ETH" {
-      ethIndex = i
-    }
-  }
+	// Print Balances
+	for i, a := range accounts {
+		//println(a.Balance)
+		if a.Currency == "USD" {
+			usdIndex = i
+		}
+		if a.Currency == "BTC" {
+			btcIndex = i
+		}
+		if a.Currency == "ETH" {
+			ethIndex = i
+		}
+	}
 }
