+++ util.go
@@ -1,0 +1,10 @@
+package main
+
+func Contains(s []float64, e float64) bool {
+	for _, a := range s {
+		if a == e {
+			return true
+		}
+	}
+	return false
+}
